requirements.txt
streamlit
pymupdf
Pillow
pandas
plotly
opencv-python-headless
numpy

